:recycle: Create addAccountTitleHOC
